Traits was moved to the concerns namespace
