+++ src/Console/Status.php
@@ -2,9 +2,9 @@
 
 namespace DragonCode\LaravelActions\Console;
 
+use DragonCode\LaravelActions\Concerns\Database;
+use DragonCode\LaravelActions\Concerns\Infoable;
 use DragonCode\LaravelActions\Constants\Names;
-use DragonCode\LaravelActions\Traits\Database;
-use DragonCode\LaravelActions\Traits\Infoable;
 use Illuminate\Database\Console\Migrations\StatusCommand as BaseCommand;
 
 class Status extends BaseCommand
