--- src/Dumper/Connections/MySql.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DragonCode\LaravelActions\Dumper\Connections;
-
-class MySql
-{
-}
