+++ src/Dumper/Schemas/SQLite.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DragonCode\LaravelActions\Dumper\Schemas;
+
+class SQLite
+{
+}
